relocate templates
