+++ template/src/{{ project_slug }}/core/paths.py
@@ -2,6 +2,15 @@ from pathlib import Path
 
 
 root: Path = Path(__file__).resolve().parent.parent.parent
+fallback_path = "default"
+
+
+def to_template_path_list(path: Path) -> list[Path]:
+    template_paths = [path]
+    template_paths.append(path / fallback_path)
+
+    return template_paths
+
 
 # Locales
 locales = root / "locales"
@@ -11,7 +20,7 @@ config_vars = root / ".copier-answers.yml"
 
 # Template paths
 templates = root / "templates"
-frontend_templates = templates / "frontend"
+frontend_templates = to_template_path_list(templates / "frontend")
 email_templates = templates / "email"
 markdown_templates = templates / "markdown"
