+++ template/templates/frontend/defaults/email_sent.html.jinja
